Core: Moved VolumeRAMPrecision to VolumeRAM, LayerRAMPrecision to LayerRAM, and BufferRAMPrecision to BufferRAM
