+++ include/inviwo/core/datastructures/image/layerram.h
@@ -36,6 +36,10 @@
 #include <inviwo/core/util/formatdispatching.h>
 #include <inviwo/core/util/glmvec.h>
 
+#include <algorithm>
+
+#include <glm/gtx/component_wise.hpp>
+
 namespace inviwo {
 
 /**
@@ -142,13 +146,315 @@ public:
     auto dispatch(Callable&& callable, Args&&... args) const -> Result;
 };
 
+/**
+ * \ingroup datastructures
+ */
+template <typename T>
+class LayerRAMPrecision : public LayerRAM {
+public:
+    using type = T;
+
+    explicit LayerRAMPrecision(size2_t dimensions = size2_t(8, 8),
+                               LayerType type = LayerType::Color,
+                               const SwizzleMask& swizzleMask = swizzlemasks::rgba,
+                               InterpolationType interpolation = InterpolationType::Linear,
+                               const Wrapping2D& wrap = wrapping2d::clampAll);
+    LayerRAMPrecision(T* data, size2_t dimensions, LayerType type = LayerType::Color,
+                      const SwizzleMask& swizzleMask = swizzlemasks::rgba,
+                      InterpolationType interpolation = InterpolationType::Linear,
+                      const Wrapping2D& wrap = wrapping2d::clampAll);
+    LayerRAMPrecision(const LayerRAMPrecision<T>& rhs);
+    LayerRAMPrecision<T>& operator=(const LayerRAMPrecision<T>& that);
+    virtual LayerRAMPrecision<T>* clone() const override;
+    virtual ~LayerRAMPrecision() = default;
+
+    T* getDataTyped();
+    const T* getDataTyped() const;
+
+    virtual void* getData() override;
+    virtual const void* getData() const override;
+    virtual void setData(void* data, size2_t dimensions) override;
+
+    /**
+     * Resize the representation to dimension. This is destructive, the data will not be
+     * preserved. Use copyRepresentationsTo to update the data.
+     */
+    virtual void setDimensions(size2_t dimensions) override;
+    const size2_t& getDimensions() const override;
+
+    virtual void setSwizzleMask(const SwizzleMask& mask) override;
+    virtual SwizzleMask getSwizzleMask() const override;
+
+    virtual void setInterpolation(InterpolationType interpolation) override;
+    virtual InterpolationType getInterpolation() const override;
+
+    virtual void setWrapping(const Wrapping2D& wrapping) override;
+    virtual Wrapping2D getWrapping() const override;
+
+    virtual double getAsDouble(const size2_t& pos) const override;
+    virtual dvec2 getAsDVec2(const size2_t& pos) const override;
+    virtual dvec3 getAsDVec3(const size2_t& pos) const override;
+    virtual dvec4 getAsDVec4(const size2_t& pos) const override;
+
+    virtual void setFromDouble(const size2_t& pos, double val) override;
+    virtual void setFromDVec2(const size2_t& pos, dvec2 val) override;
+    virtual void setFromDVec3(const size2_t& pos, dvec3 val) override;
+    virtual void setFromDVec4(const size2_t& pos, dvec4 val) override;
+
+    virtual double getAsNormalizedDouble(const size2_t& pos) const override;
+    virtual dvec2 getAsNormalizedDVec2(const size2_t& pos) const override;
+    virtual dvec3 getAsNormalizedDVec3(const size2_t& pos) const override;
+    virtual dvec4 getAsNormalizedDVec4(const size2_t& pos) const override;
+
+    virtual void setFromNormalizedDouble(const size2_t& pos, double val) override;
+    virtual void setFromNormalizedDVec2(const size2_t& pos, dvec2 val) override;
+    virtual void setFromNormalizedDVec3(const size2_t& pos, dvec3 val) override;
+    virtual void setFromNormalizedDVec4(const size2_t& pos, dvec4 val) override;
+
+private:
+    size2_t dimensions_;
+    std::unique_ptr<T[]> data_;
+    SwizzleMask swizzleMask_;
+    InterpolationType interpolation_;
+    Wrapping2D wrapping_;
+};
+
+/**
+ * Factory for layers.
+ * Creates an LayerRAM with data type specified by format.
+ *
+ * @param dimensions of layer to create.
+ * @param type of layer to create.
+ * @param format of layer to create.
+ * @param swizzleMask used in for the layer, defaults to RGB-alpha
+ * @param interpolation method to use.
+ * @param wrapping method to use.
+ * @return nullptr if no valid format was specified.
+ */
+IVW_CORE_API std::shared_ptr<LayerRAM> createLayerRAM(
+    const size2_t& dimensions, LayerType type, const DataFormatBase* format,
+    const SwizzleMask& swizzleMask = swizzlemasks::rgba,
+    InterpolationType interpolation = InterpolationType::Linear,
+    const Wrapping2D& wrapping = wrapping2d::clampAll);
+
+template <typename T>
+LayerRAMPrecision<T>::LayerRAMPrecision(size2_t dimensions, LayerType type,
+                                        const SwizzleMask& swizzleMask,
+                                        InterpolationType interpolation, const Wrapping2D& wrapping)
+    : LayerRAM(type, DataFormat<T>::get())
+    , dimensions_(dimensions)
+    , data_(new T[dimensions_.x * dimensions_.y]())
+    , swizzleMask_(swizzleMask)
+    , interpolation_{interpolation}
+    , wrapping_{wrapping} {
+    std::fill(data_.get(), data_.get() + glm::compMul(dimensions_),
+              (type == LayerType::Depth) ? T{1} : T{0});
+}
+
+template <typename T>
+LayerRAMPrecision<T>::LayerRAMPrecision(T* data, size2_t dimensions, LayerType type,
+                                        const SwizzleMask& swizzleMask,
+                                        InterpolationType interpolation, const Wrapping2D& wrapping)
+    : LayerRAM(type, DataFormat<T>::get())
+    , dimensions_(dimensions)
+    , data_(data ? data : new T[dimensions_.x * dimensions_.y]())
+    , swizzleMask_(swizzleMask)
+    , interpolation_{interpolation}
+    , wrapping_{wrapping} {
+    if (!data) {
+        std::fill(data_.get(), data_.get() + glm::compMul(dimensions_),
+                  (type == LayerType::Depth) ? T{1} : T{0});
+    }
+}
+
+template <typename T>
+LayerRAMPrecision<T>::LayerRAMPrecision(const LayerRAMPrecision<T>& rhs)
+    : LayerRAM(rhs)
+    , dimensions_(rhs.dimensions_)
+    , data_(new T[dimensions_.x * dimensions_.y])
+    , swizzleMask_(rhs.swizzleMask_)
+    , interpolation_{rhs.interpolation_}
+    , wrapping_{rhs.wrapping_} {
+    std::memcpy(data_.get(), rhs.data_.get(), dimensions_.x * dimensions_.y * sizeof(T));
+}
+
+template <typename T>
+LayerRAMPrecision<T>& LayerRAMPrecision<T>::operator=(const LayerRAMPrecision<T>& that) {
+    if (this != &that) {
+        LayerRAM::operator=(that);
+
+        const auto dim = that.dimensions_;
+        auto data = std::make_unique<T[]>(dim.x * dim.y);
+        std::memcpy(data.get(), that.data_.get(), dim.x * dim.y * sizeof(T));
+        data_.swap(data);
+
+        dimensions_ = that.dimensions_;
+        swizzleMask_ = that.swizzleMask_;
+        interpolation_ = that.interpolation_;
+        wrapping_ = that.wrapping_;
+    }
+    return *this;
+}
+
+template <typename T>
+LayerRAMPrecision<T>* LayerRAMPrecision<T>::clone() const {
+    return new LayerRAMPrecision<T>(*this);
+}
+
+template <typename T>
+T* inviwo::LayerRAMPrecision<T>::getDataTyped() {
+    return data_.get();
+}
+
+template <typename T>
+const T* inviwo::LayerRAMPrecision<T>::getDataTyped() const {
+    return data_.get();
+}
+
+template <typename T>
+void* LayerRAMPrecision<T>::getData() {
+    return data_.get();
+}
+template <typename T>
+const void* LayerRAMPrecision<T>::getData() const {
+    return const_cast<const T*>(data_.get());
+}
+
+template <typename T>
+void inviwo::LayerRAMPrecision<T>::setData(void* d, size2_t dimensions) {
+    std::unique_ptr<T[]> data(static_cast<T*>(d));
+    data_.swap(data);
+    std::swap(dimensions_, dimensions);
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setDimensions(size2_t dimensions) {
+    if (dimensions != dimensions_) {
+        auto data = std::make_unique<T[]>(dimensions.x * dimensions.y);
+        data_.swap(data);
+        std::swap(dimensions, dimensions_);
+    }
+}
+
+template <typename T>
+const size2_t& LayerRAMPrecision<T>::getDimensions() const {
+    return dimensions_;
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setSwizzleMask(const SwizzleMask& mask) {
+    swizzleMask_ = mask;
+}
+
+template <typename T>
+SwizzleMask LayerRAMPrecision<T>::getSwizzleMask() const {
+    return swizzleMask_;
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setInterpolation(InterpolationType interpolation) {
+    interpolation_ = interpolation;
+}
+
+template <typename T>
+InterpolationType LayerRAMPrecision<T>::getInterpolation() const {
+    return interpolation_;
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setWrapping(const Wrapping2D& wrapping) {
+    wrapping_ = wrapping;
+}
+
+template <typename T>
+Wrapping2D LayerRAMPrecision<T>::getWrapping() const {
+    return wrapping_;
+}
+
+template <typename T>
+double LayerRAMPrecision<T>::getAsDouble(const size2_t& pos) const {
+    return util::glm_convert<double>(data_[posToIndex(pos, dimensions_)]);
+}
+
+template <typename T>
+dvec2 LayerRAMPrecision<T>::getAsDVec2(const size2_t& pos) const {
+    return util::glm_convert<dvec2>(data_[posToIndex(pos, dimensions_)]);
+}
+
+template <typename T>
+dvec3 LayerRAMPrecision<T>::getAsDVec3(const size2_t& pos) const {
+    return util::glm_convert<dvec3>(data_[posToIndex(pos, dimensions_)]);
+}
+
+template <typename T>
+dvec4 LayerRAMPrecision<T>::getAsDVec4(const size2_t& pos) const {
+    return util::glm_convert<dvec4>(data_[posToIndex(pos, dimensions_)]);
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setFromDouble(const size2_t& pos, double val) {
+    data_[posToIndex(pos, dimensions_)] = util::glm_convert<T>(val);
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setFromDVec2(const size2_t& pos, dvec2 val) {
+    data_[posToIndex(pos, dimensions_)] = util::glm_convert<T>(val);
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setFromDVec3(const size2_t& pos, dvec3 val) {
+    data_[posToIndex(pos, dimensions_)] = util::glm_convert<T>(val);
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setFromDVec4(const size2_t& pos, dvec4 val) {
+    data_[posToIndex(pos, dimensions_)] = util::glm_convert<T>(val);
+}
+
+template <typename T>
+double LayerRAMPrecision<T>::getAsNormalizedDouble(const size2_t& pos) const {
+    return util::glm_convert_normalized<double>(data_[posToIndex(pos, dimensions_)]);
+}
+
+template <typename T>
+dvec2 LayerRAMPrecision<T>::getAsNormalizedDVec2(const size2_t& pos) const {
+    return util::glm_convert_normalized<dvec2>(data_[posToIndex(pos, dimensions_)]);
+}
+
+template <typename T>
+dvec3 LayerRAMPrecision<T>::getAsNormalizedDVec3(const size2_t& pos) const {
+    return util::glm_convert_normalized<dvec3>(data_[posToIndex(pos, dimensions_)]);
+}
+
+template <typename T>
+dvec4 LayerRAMPrecision<T>::getAsNormalizedDVec4(const size2_t& pos) const {
+    return util::glm_convert_normalized<dvec4>(data_[posToIndex(pos, dimensions_)]);
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setFromNormalizedDouble(const size2_t& pos, double val) {
+    data_[posToIndex(pos, dimensions_)] = util::glm_convert_normalized<T>(val);
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setFromNormalizedDVec2(const size2_t& pos, dvec2 val) {
+    data_[posToIndex(pos, dimensions_)] = util::glm_convert_normalized<T>(val);
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setFromNormalizedDVec3(const size2_t& pos, dvec3 val) {
+    data_[posToIndex(pos, dimensions_)] = util::glm_convert_normalized<T>(val);
+}
+
+template <typename T>
+void LayerRAMPrecision<T>::setFromNormalizedDVec4(const size2_t& pos, dvec4 val) {
+    data_[posToIndex(pos, dimensions_)] = util::glm_convert_normalized<T>(val);
+}
+
 size_t inline LayerRAM::posToIndex(const size2_t& pos, const size2_t& dim) {
     IVW_ASSERT((pos.x < dim.x) && (pos.y < dim.y), "posToIndex: position out of bounds");
     return pos.x + (pos.y * dim.x);
 }
-
-template <typename T>
-class LayerRAMPrecision;
 
 namespace detail {
 struct LayerRamDispatcher {
