+++ modules/base/src/processors/imagestackvolumesource.cpp
@@ -43,12 +43,15 @@
 #include <inviwo/core/util/vectoroperations.h>
 #include <inviwo/core/util/zip.h>
 #include <inviwo/core/util/raiiutils.h>
+#include <inviwo/core/util/glm.h>
 #include <inviwo/core/io/datareaderexception.h>
 
 #include <algorithm>
 
 #include <fmt/format.h>
 #include <fmt/ostream.h>
+
+
 
 namespace inviwo {
